chore: extract out async_unsafe decorator into a helper file
