+++ packages/django-cf/django_cf/db/base_engine.py
@@ -43,7 +43,8 @@ from django.db.models.functions import (
     TruncYear,
 )
 from django.db.models.sql.compiler import SQLCompiler
-from django.utils import asyncio as _django_asyncio
+
+from ._async_unsafe import _unwrap_async_unsafe
 
 
 def replace_date_trunc_in_sql(sql):
@@ -479,35 +480,6 @@ class CFSQLCompiler(SQLCompiler):
 
         sql = templates.get(kind, field_sql)
         return sql, params
-
-
-def _async_unsafe_probe():
-    pass
-
-
-# Capture async_unsafe function so we don't accidentally unwrap something else
-_ASYNC_UNSAFE_WRAPPER_CODE = _django_asyncio.async_unsafe(_async_unsafe_probe).__code__
-
-
-def _unwrap_async_unsafe(method):
-    """
-    Unwrap Django's async_unsafe decorator to get the original method.
-
-    This is needed because Django does not allow calling database operations
-    from within an async context, but Python workers always run in an async context.
-
-    All the database backends that django-cf provides are async-safe so we need
-    to unwrap the async_unsafe decorator to allow calling database operations
-    from within an async context.
-    """
-    method_code = getattr(method, "__code__", None)
-    is_async_unsafe = method_code is _ASYNC_UNSAFE_WRAPPER_CODE
-    wrapped = getattr(method, "__wrapped__", None)
-
-    if is_async_unsafe:
-        return wrapped
-
-    return method
 
 
 class CFDatabaseWrapper(SQLiteDatabaseWrapper):
